Update README.md
In linguistics, a compound word is two or more individual words used together to create a new word with its own individual meaning. Because there are so many compound words in English, we may not actually recognize them as being compounds.

Firstly we create a class named LongestCompoundWord.
Then we take a new LinkedList.
Then we used trie data structure for efficient retrieval of a key in a huge word set.
Then we take a pair of word and its remaining suffix.
Then we defined longest and 2nd longest compound word.
if no suffixes found, which means no prefixes found discard the pair and check the next pair.
Then we print the given word by user.
No suffix, means it is a compound word check if the compound word is the longest if it is update both longest and second longest words records.
Then it will search if it is longest compound word or not?
Finall if it is longest compound word then it will print.